requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
playwright
beautifulsoup4
pyyaml
lxml